# Alerting: Optimize rule status API when there is no limit specified